chore: use forked tar-rs suitable for multi-step tar parsing
The bug happened because `tar-rs` automatically updates its inner
position to the one returned by the `Seek` implementation, which didn't
work quite well with my nested tar layer parsing logic and caused an
underflow (because `Archive` is not prepared for someone to parse a tar
in multiple steps).
There is actually no need to update the position that way, as we can
just add the amount by which we seek directly.